omega_executor/requirements.txt
flask
requests
pybit
python-dotenv
pandas
scikit-learn
joblib
schedule
python-telegram-bot
matplotlib
ccxt
python-dotenv
flask